CODERUB: Move ShouldRetrieveAllStudentsViewAsync to another partial class
