+++ OtripleS.Portal.Web.Tests.Unit/Services/StudentViews/StudentViewServiceTests.Logic.cs
@@ -122,66 +122,5 @@ namespace OtripleS.Portal.Web.Tests.Unit.Services.StudentViews
             this.studentServiceMock.VerifyNoOtherCalls();
             this.loggingBrokerMock.VerifyNoOtherCalls();
         }
-
-        [Fact]
-        public async Task ShouldRetrieveAllStudentsViewAsync()
-        {
-            var randomUserId = Guid.NewGuid();
-            DateTimeOffset randomDateTime = GetRandomDate();
-
-            dynamic dynamicStudentProperties =
-                CreateRandomStudentViewProperties(
-                    auditDates: randomDateTime,
-                    auditIds: randomUserId);
-
-            var student = new Student
-            {
-                Id = dynamicStudentProperties.Id,
-                UserId = dynamicStudentProperties.UserId,
-                IdentityNumber = dynamicStudentProperties.IdentityNumber,
-                FirstName = dynamicStudentProperties.FirstName,
-                MiddleName = dynamicStudentProperties.MiddleName,
-                LastName = dynamicStudentProperties.LastName,
-                Gender = dynamicStudentProperties.Gender,
-                BirthDate = dynamicStudentProperties.BirthDate,
-                CreatedDate = randomDateTime,
-                UpdatedDate = randomDateTime,
-                CreatedBy = randomUserId,
-                UpdatedBy = randomUserId
-            };
-
-            var studentView = new StudentView
-            {
-                IdentityNumber = dynamicStudentProperties.IdentityNumber,
-                FirstName = dynamicStudentProperties.FirstName,
-                MiddleName = dynamicStudentProperties.MiddleName,
-                LastName = dynamicStudentProperties.LastName,
-                Gender = (StudentViewGender)dynamicStudentProperties.Gender,
-                BirthDate = dynamicStudentProperties.BirthDate
-            };
-
-            var randomStudents = new List<Student> { student };
-            var retrievedServiceStudents = randomStudents;
-            var expectedStudentViews = new List<StudentView> { studentView };
-
-            this.studentServiceMock.Setup(service =>
-                service.RetrieveAllStudentsAsync())
-                    .ReturnsAsync(retrievedServiceStudents);
-
-            List<StudentView> retrievedStudentViews =
-                await this.studentViewService.RetrieveAllStudentsViewAsync();
-
-            retrievedStudentViews.Should().BeEquivalentTo(expectedStudentViews);
-
-            this.studentServiceMock.Verify(service =>
-                service.RetrieveAllStudentsAsync(),
-                    Times.Once());
-
-            this.studentServiceMock.VerifyNoOtherCalls(); 
-            this.navigationBrokerMock.VerifyNoOtherCalls();
-            this.dateTimeBrokerMock.VerifyNoOtherCalls();
-            this.userServiceMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-        }
     }
 }
